fix review comments - 3
